NoiseCalc - changes to helicopter tasks and new dropdown for choice of tasks.
git-tfs-id: [http://tfs.statoil.net:8080/tfs/TFS2005]$/Statoil.SmaJobbSentralen.RetailSE/Source/NoiseCalculator/Source/Main;C27705

+++ NoiseCalculator.Domain/Entities/SelectedTask.cs
@@ -8,7 +8,7 @@ namespace NoiseCalculator.Domain.Entities
         public virtual string Title { get; set; }
         public virtual string Role { get; set; }
         public virtual string NoiseProtection { get; set; }
-        public virtual int NoiseLevel { get; set; }
+        public virtual decimal NoiseLevel { get; set; }
         public virtual int Hours { get; set; }
         public virtual int Minutes { get; set; }
         public virtual int Percentage { get; set; }
