+++ NoiseCalculator.UI.Web/Areas/Admin/ViewModels/GenericTask/TaskListItemViewModel.cs
@@ -6,7 +6,7 @@
         public string Title { get; set; }
         public string Role { get; set; }
         public string NoiseProtection { get; set; }
-        public int NoiseLevelGuideline { get; set; }
+        public decimal NoiseLevelGuideline { get; set; }
         public int AllowedExposureMinutes { get; set; }
         public int ButtonPressed { get; set; }
         public int BackgroundNoise { get; set; }
